requirements.txt
# requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.35.0
streamlit>=1.28.0
plotly>=5.14.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0
prophet>=1.1.0
arch>=6.0.0
torchvision>=0.15.0
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
docker>=6.1.0